refactor: rename "Info Card" to "Card Info" and move him to the customs folder
