--- src/components/global/Cards/InfoCard/styles.sass
@@ -1,21 +1,0 @@
-@use "../../../../assets/sass/variables/bootstrap" as *
-@use "../../../../assets/sass/mixins/typography" as *
-
-// Component styling
-.info-card
-  display: flex
-  align-items: center
-  justify-content: center
-  gap: 12px
-  background-color: $color-primary-1
-  min-width: 110px
-  padding: 6px 0 4px
-  border-radius: 4px
-  // box-shadow: $shadow-primary-1
-
-  &__icon
-    font-size: 1.25rem
-    color: $color-secondary-2
-
-  &__label
-    @include font-custom(10, 14, left, 300, "Outfit", $color-secondary-2)
